pyproject.toml
[project]
name = "agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.15.0",
    "fastapi>=0.115.12",
    "google-genai>=1.19.0",
    "huggingface-hub>=0.32.4",
    "lancedb>=0.23.0",
    "pandas>=2.3.0",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
]
